requirements-pi.txt
Pillow
websockets
